Добавлвена model ViewStats Исправлено не правильное поведение sql запросов Убрано лишнее
