+++ stats-service/client/src/main/java/ru/practicum/BaseClient.java
