+++ stats-service/client/src/main/java/ru/practicum/StatsClient.java
