+++ stats-service/client/src/main/resources/application.properties
