+++ stats-service/dto/src/main/java/ru/practicum/dto/EndpointHitDto.java
