+++ stats-service/dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -1,0 +1,16 @@
+package ru.practicum.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ViewStatsDto {
+    private String app;//название сервиса
+    private String uri;//URI сервиса
+    private Long hits;//колличество просмотров
+}
