+++ stats-service/service/.gitignore
