+++ stats-service/service/Dockerfile
