+++ stats-service/service/src/main/java/ru/practicum/StatsServiceApp.java
