+++ stats-service/service/src/main/java/ru/practicum/expeptions/ErrorResponse.java
