+++ stats-service/service/src/main/java/ru/practicum/mapper/StatsMapper.java
@@ -1,0 +1,30 @@
+package ru.practicum.mapper;
+
+import ru.practicum.dto.EndpointHitDto;
+import ru.practicum.dto.ViewStatsDto;
+import ru.practicum.model.Stats;
+import ru.practicum.model.ViewStats;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class StatsMapper {
+    private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+
+    public static Stats toStats(EndpointHitDto endpointHitDto) {
+        return Stats.builder()
+                .app(endpointHitDto.getApp())
+                .uri(endpointHitDto.getUri())
+                .ip(endpointHitDto.getIp())
+                .timestamp(LocalDateTime.parse(endpointHitDto.getTimestamp(), formatter))
+                .build();
+    }
+
+    public static ViewStatsDto toViewStatsDto(ViewStats viewStats) {
+        return ViewStatsDto.builder()
+                .app(viewStats.getApp())
+                .uri(viewStats.getUri())
+                .hits(viewStats.getHits())
+                .build();
+    }
+}
