+++ stats-service/service/src/main/java/ru/practicum/model/Stats.java
@@ -1,0 +1,31 @@
+package ru.practicum.model;
+
+import lombok.*;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "STATS", schema = "public")
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Stats {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long id;
+
+    @Column(name = "app_name", nullable = false)
+    String app;
+
+    @Column
+    String uri;
+
+    @Column(name = "user_ip", length = 15)
+    String ip;
+
+    @Column(name = "created")
+    LocalDateTime timestamp;
+}
