+++ stats-service/service/src/main/java/ru/practicum/service/StatsService.java
