+++ stats-service/service/src/main/resources/application.properties
