+++ stats-service/service/src/main/resources/schema.sql
