--- statsService/service/src/main/java/ru/practicum/mapper/StatsMapper.java
@@ -1,29 +1,0 @@
-package ru.practicum.mapper;
-
-import ru.practicum.dto.EndpointHitDto;
-import ru.practicum.dto.ViewStatsDto;
-import ru.practicum.model.Stats;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class StatsMapper {
-    private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-
-    public static Stats toStats(EndpointHitDto endpointHitDto) {
-        return Stats.builder()
-                .app(endpointHitDto.getApp())
-                .uri(endpointHitDto.getUri())
-                .ip(endpointHitDto.getIp())
-                .timestamp(LocalDateTime.parse(endpointHitDto.getTimestamp(), formatter))
-                .build();
-    }
-
-    public static ViewStatsDto toViewStatsDto(Stats stats, Long hits) {
-        return ViewStatsDto.builder()
-                .app(stats.getApp())
-                .uri(stats.getUri())
-                .hits(hits)
-                .build();
-    }
-}
