--- statsService/service/src/main/java/ru/practicum/repository/StatsRepository.java
@@ -1,47 +1,0 @@
-package ru.practicum.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import ru.practicum.dto.ViewStatsDto;
-import ru.practicum.model.Stats;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface StatsRepository extends JpaRepository<Stats, Long> {
-    @Query("SELECT new ru.practicum.dto.ViewStatsDto(s.app, s.uri, COUNT(DISTINCT s.ip)) " +
-            "FROM Stats AS s " +
-            "WHERE s.timestamp BETWEEN ?1 AND ?2 " +
-            "GROUP BY s.app, s.uri " +
-            "ORDER BY COUNT(DISTINCT s.ip) DESC")
-    List<ViewStatsDto> getAllWithUniq(LocalDateTime start, LocalDateTime end);
-
-    @Query("SELECT COUNT(s.ip) " +
-            "FROM Stats AS s " +
-            "WHERE s.timestamp BETWEEN ?1 AND ?2 AND s.ip = ?3 " +
-            "GROUP BY s.app, s.uri " +
-            "ORDER BY COUNT(s.ip)")
-    Long getStatsCount(LocalDateTime start, LocalDateTime end, String ip);
-
-    @Query("SELECT new ru.practicum.dto.ViewStatsDto(s.app, s.uri, COUNT(s.ip)) " +
-            "FROM Stats AS s " +
-            "WHERE s.timestamp BETWEEN ?1 AND ?2 " +
-            "GROUP BY s.app, s.uri " +
-            "ORDER BY COUNT(s.ip) DESC")
-    List<ViewStatsDto> getAll(LocalDateTime start, LocalDateTime end);
-
-    @Query("SELECT new ru.practicum.dto.ViewStatsDto(s.app, s.uri, COUNT(DISTINCT s.ip)) " +
-            "FROM Stats AS s " +
-            "WHERE s.timestamp BETWEEN ?1 AND ?2 AND s.uri IN (?3)" +
-            "GROUP BY s.app, s.uri " +
-            "ORDER BY COUNT(DISTINCT s.ip) DESC")
-    List<ViewStatsDto> getAllWithUrisAndUniq(LocalDateTime start, LocalDateTime end, List<String> uris);
-
-    @Query("SELECT new ru.practicum.dto.ViewStatsDto(s.app, s.uri, COUNT(s.ip)) " +
-            "FROM Stats AS s " +
-            "WHERE s.timestamp BETWEEN ?1 AND ?2 " +
-            "AND s.uri IN (?3) " +
-            "GROUP BY s.app, s.uri " +
-            "ORDER BY COUNT(s.ip) DESC")
-    List<ViewStatsDto> getAllWithUri(LocalDateTime start, LocalDateTime end, List<String> uris);
-}
